News from the inn?
Updates the NewsFromTheInn on the game state when we check for corruption
as a Clairvoyant.

This currently doesn't handle the case where the clairvoyant checks but
the Bard/Innkeeper is subsequently killed or made into a minion before
morning.

Also adds some utility methods to the game state to extract a player
given a role, or to check if a role is still in play.