Handle adding players through interface
Updates the start page to allow players to be added with a role.
As roles are dished out, they are removed from the available roles to
assign subsequent players.

Players can only be entered when they have a non-empty name

Game is stored as part of a singleton, available statically.