Use filters for playersAlive and playersDead
I planned when writing the player that the player.alive boolean would be
updated when a player dies, then forgot about it.

Now, we have a list of allPlayers. state.playersAlive is simply a filtered
version of allPlayers. Killing a player only means setting player.alive = NO,
and the lists update themselves.