Fill in "Clairvoyant check" view
Finally worked out how to display the Clairvoyant view without obscuring
the Tab bar controller!

Here we add a selection table to check a single player for Corruption,
then reveal the results of the check. Currently this does not handle the
case where the clairvoyant is not in play.

We introduce Segues between controllers, allowing us to choose by
circumstance which view controller we need to display.